John 2018 10.12
I made a bunch of changes to the editor.java file just to get rid of the compiling issues and mass use the sample connectors given. I was able to get the select queries working for the posts table. But when I compile the jsps are not updating, need to figure out why.